+++ src/matchbox/client/dags.py
@@ -17,7 +17,7 @@ from matchbox.client.models.dedupers.base import Deduper
 from matchbox.client.models.linkers.base import Linker
 from matchbox.client.models.models import make_model
 from matchbox.client.results import Results
-from matchbox.common.dtos import ResolutionName, SourceResolutionName
+from matchbox.common.graph import ResolutionName, SourceResolutionName
 from matchbox.common.logging import logger
 from matchbox.common.sources import RelationalDBLocation, SourceConfig, SourceField
