+++ test/common/factories/test_entity_factory.py
@@ -4,7 +4,7 @@ import pyarrow as pa
 import pytest
 from faker import Faker
 
-from matchbox.common.dtos import DataTypes, SourceResolutionName
+from matchbox.common.dtos import DataTypes
 from matchbox.common.factories.entities import (
     ClusterEntity,
     EntityReference,
@@ -14,6 +14,7 @@ from matchbox.common.factories.entities import (
     generate_entities,
     probabilities_to_results_entities,
 )
+from matchbox.common.graph import SourceResolutionName
 
 
 def make_cluster_entity(id: int, *args) -> ClusterEntity:
